+++ CHANGELOG.md
@@ -3,6 +3,7 @@ LUYA CHANGELOG
 
 1.0.0-alpha9 (in progress)
 --------------------------
+- removed `$links->activeLink` and replaced with `$links->activeUrl` in `#196`.
 
 
 1.0.0-alpha8 (30. Jun 2015)